Update README.md (#60)
Changes to the README.md based on comments from the doc team review.

We have to decide on the compromise between staying consistent with the
template repo and being factually correct and clear. I can revert both
of the edits if you prefer to stay with the original, but it seems that
only a few repos stick with the exact same text as the template, so I'm
tempted to go with clarity/accuracy in this repo over consistency with a
few other Ansys repos, but with only the minimal changes.